+++ NEWS.md
@@ -1,3 +1,53 @@
+<<<<<<< HEAD
+=======
+# SmartEDA 0.3.7
+## New Features
+* Added function `ExpTwoPlots` to graph from same variable when Target=NULL vs. when Target = categorical variable (binary or multi-class variable)
+
+## Enhancements
+* Added "fun" option to add custom statistics on data dictionary output in `ExpData`
+
+# SmartEDA 0.3.6
+## Bug fixes
+* Problem with date format. Fixed a bug in `ExpData()`, function is now supporting for all types of columns. Previously this function were not supporting to data types POSIXlt or date format
+* Fixed `ExpNumViz()` code in the README file. Previously it contains arguments that are not (anymore) present in the function
+
+## Enhancements
+* Added new outcome for `ExpOutliers()` called outlier index value. Outlier index for both upper and lower outliers
+* Column name changed from returned output of `ExpData()` type 1 - "Obs" to new name "Value"
+* Column name changed from returned output of `ExpData()` type 2 - "S.no" to new name "Index", "% of Missing" as "Per_of_Missing"
+* Corrected few spell mistakes from `ExpData()` help file
+* Corrected output readings from `ExpData()` - Unqiue variable replaced as identifier variables
+
+# SmartEDA 0.3.5
+## Bug fixes
+* Fixed a bug in `ExpCatStat`, Corrected odds ratio calculation for mutliclass
+* Fixed a bug in `ExpData`, for Type = 2 changed output variable structure - previously all columns are factors and now changed to character and interger values. 
+
+## Enhancements
+* Removed asterik mark from factor and date variables in `ExpData` type = 2
+* Added unique variable count in `ExpData` type = 1
+* Missing value percentage will be calculated based on NA value and no. of blanks
+* Added Target column in `ExpCatStat` output where stat = "IV"
+
+# SmartEDA 0.3.4
+## Enhancements
+* Added mutliple standard devitaion options to treat oultier value in `ExpOutliers`
+
+## Bug fixes
+* Fixed error Centos 7 MRO, removed grDevices hcl.colours dependencies
+
+# SmartEDA 0.3.3
+## New Features
+* Added function `ExpOutliers` to run univariate outlier analysis
+
+# SmartEDA 0.3.2
+## Enhancements
+* Added 'scatter' option to plot the correlation plot between all the numeric variables in a dataframe `ExpNumViz`
+* Added 'theme' option to customise the graph theme in `ExpReport`
+* Changed input parameter name from 'gp' to 'target' in `ExpNumViz`
+
+>>>>>>> master
 # SmartEDA 0.3.1
 ## Enhancements
 * Added 'bins', 'plot', 'round' and 'top' options to plot bar graph in `ExpCatStat`
@@ -9,7 +59,10 @@
 ## Bug fixes
 * Fixed a formula issues on odds calculation in `ExpCatStat`
 
+<<<<<<< HEAD
 
+=======
+>>>>>>> master
 # SmartEDA 0.3.0
 ## New Features
 * Added function `ExpOutQQ` to plot Quantile-Quantile Plots for outlier checking